refactor(engine): close execution service compatibility paths
Route stdin root bytes through the owned execution world and delete the
optional filesystem child runner so CLI and ARM share one immutable
admission boundary. Add structural extinction ratchets and refresh
the W04 architecture evidence.
