+++ crates/nika-cli/src/verbs/run/tests.rs
@@ -29,7 +29,6 @@ fn registry_run_refusal_keeps_copy_guidance_and_never_names_cache_as_fixable() {
     let cache_path = path.to_string_lossy().into_owned();
     let (source, _wf, report) = super::provenance::capture_checked_source(
         &cache_path,
-        None,
         Some(nika_display::check_render::RepairTarget::RegistryArtifact),
         false,
     )
