+++ crates/nika-cli/src/verbs/run/thread.rs
@@ -33,7 +33,6 @@ pub(crate) fn run_in_thread(file: &str, theme: Theme) -> ThreadRun {
         false,
         true,
         None,
-        None,
     );
     ThreadRun {
         answer: verdict
